IMemoryHolder now returns a Span instead of a IntPtr Things may be broken, especially ImageReadHelpers

+++ src/StbImageSharp/ImageRead.ReadContext.cs
@@ -150,7 +150,7 @@ namespace StbSharp
             public byte ReadByte()
             {
                 if (Data < DataEnd)
-                    return (byte)*Data++;
+                    return *Data++;
 
                 if (ReadFromCallbacks)
                 {
